requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
numpy==1.26.0
packaging==23.2
pickleshare==0.7.5
Pillow==10.0.1
scikit-learn==1.3.1
scipy==1.11.3
threadpoolctl==3.2.0
Werkzeug==3.0.0
